Major re-write. imap -> webhooks. zbar.
Previously the server was based around the use of an imap server, but this isn't really that scalable, and it's possible that the imap host wouldn't like this kind of traffic.

Also in this commit is some major refactoring to remove the imap and move to the webhook system.

Also in this commit is the integration of zbar. zxing is a nice java library, but unfortunately it does not always process qr codes. zbar is much more reliable in this sense.

There are also some additions to make the server work on heroku.